requirements.txt
langchain==0.0.250
pypdf==3.15.2
python-dotenv==1.0.0
streamlit==1.25.0
chromadb==0.4.6
faiss-cpu==1.7.4
tiktoken==0.4.0
altair<5
openai